killing the 'with' statement
with was initially an experiment to see how well the interpreter and
runtime would be able to accomodate it. it wasn't bad, but it always
had limited utility, and i never had the chance to integrate it into
blockly. now it's getting more difficult to keep its simple design
while evolving the interpreter. so, goodbye for now.