interpreter cleanup
proof of concept, pass a mutable frame to node methods to drastically
simplify maintaining interpreter state. the run loop to make it work
is more complex, since it has to deal with flow instructions after the
fact, but the node logic readability is worth it.

this takes advantage of the new Record implementation in immutable 4,
which makes field access way faster.