change namespace semantics
instead of a single namespace stack to search, maintain a single global namespace,
and a local namespace stack. when looking up a var, look in the top local ns first,
and then the global ns. when setting, always use the local ns if local=true. simple.

i think this makes the scoping rules a little less surprising for beginners. it's
either local, or it's global, no temporal scoping to worry about.

also cleanup/renames.